working guessthenumber

+++ Summatives/GuessTheNumber/GuessTheNumberREST/src/main/java/com/mc/guessthenumberrest/controllers/GuessTheNumberRESTController.java
@@ -7,19 +7,19 @@ package com.mc.guessthenumberrest.controllers;
 
 import com.mc.guessthenumberrest.models.Game;
 import com.mc.guessthenumberrest.models.Round;
-import com.mc.guessthenumberrest.data.GuessTheNumberRESTServiceLayer;
-import static java.lang.Boolean.parseBoolean;
+import com.mc.guessthenumberrest.service.GuessTheNumberRESTServiceLayer;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
@@ -35,18 +35,19 @@ public class GuessTheNumberRESTController {
 
     private final GuessTheNumberRESTServiceLayer service;
 
+    @Autowired
     public GuessTheNumberRESTController(GuessTheNumberRESTServiceLayer service) {
         this.service = service;
     }
 
     @GetMapping
-    public List<Game> all() {       //working
+    public List<Game> all() {
         return service.getAllGames();
     }
 
     @PostMapping("/begin")
     @ResponseStatus(HttpStatus.CREATED)
-    public int begin() {                //working
+    public int begin() {
         String answer = service.createAnswer();
         Game newGame = new Game();
         newGame.setGameAnswer(answer);
@@ -57,25 +58,23 @@ public class GuessTheNumberRESTController {
 
     @PostMapping("/guess")
     @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<Round> guess(String roundGuess,  int gameId) {
-
-        HashMap<String, String> roundResults = new HashMap();
+    public ResponseEntity<Round> guess(@RequestBody Round round) {
         Round currentRound = new Round();
-        
-       Game currentGame = service.getCurrentGameById(gameId);
 
-        boolean containsCorrectDigits = service.containsFourDigits(roundGuess);
+        Game currentGame = service.getCurrentGameById(round.getGameId());
 
-        boolean duplicate = service.hasDuplicateDigits(Integer.parseInt(roundGuess));
+        boolean containsCorrectDigits = service.containsFourDigits(round.getRoundGuess());
+
+        boolean duplicate = service.hasDuplicateDigits(Integer.parseInt(round.getRoundGuess()));
 
         if (duplicate == false && containsCorrectDigits == true) {
-            roundResults = service.playRound(roundGuess, currentGame.getGameAnswer());
+            HashMap<String, String> roundResults = service.playRound(round.getRoundGuess(), currentGame.getGameAnswer());
 
             currentGame.setStatus(roundResults.values().toString());
-            currentGame.setId(gameId);
+            currentGame.setId(currentGame.getId());
             currentGame.setGameAnswer(currentGame.getGameAnswer());
-            currentRound.setGame(currentGame);
-            currentRound.setRoundGuess(roundGuess);
+            currentRound.setGameId(round.getGameId());
+            currentRound.setRoundGuess(round.getRoundGuess());
             currentRound.setTimeOfGuess(LocalDateTime.now());
             currentRound.setResultOfGuess(roundResults.keySet().toString());
 
@@ -86,9 +85,8 @@ public class GuessTheNumberRESTController {
         return new ResponseEntity(null, HttpStatus.NOT_FOUND);
     }
 
-    @GetMapping("game/{id}") //working
-    public ResponseEntity<Game> findGameById(@PathVariable int id
-    ) {
+    @GetMapping("game/{id}")
+    public ResponseEntity<Game> findGameById(@PathVariable int id) {
         Game result = service.getGameById(id);
         if (result == null) {
             return new ResponseEntity(null, HttpStatus.NOT_FOUND);
@@ -96,30 +94,31 @@ public class GuessTheNumberRESTController {
         return ResponseEntity.ok(result);
     }
 
-    @GetMapping("rounds/{id}")  //working
-    public List<Round> findRoundsForGameByGameId(@PathVariable int id){
-   
+    @GetMapping("rounds/{id}")
+    public List<Round> findRoundsForGameByGameId(@PathVariable int id) {
+
         Game game = service.getGameById(id);
         if (game == null) {
             throw new ResponseStatusException(
                     HttpStatus.NOT_FOUND, "Id not found"
             );
         }
-        List result = service.getRoundsForGame(game);//must sort by time
-        return result;
+//        List result = service.getRoundsForGame(game);//must sort by time
+        return service.getRoundsForGame(game);//must sort by time
     }
-    
-    @DeleteMapping("/{id}") //working 
-    
+
+    @DeleteMapping("/{id}")
+
     public ResponseEntity deleteGameById(@PathVariable int id) {
-        List<Round> RoundList = new ArrayList<Round>();
+//        List<Round> RoundList = new ArrayList<Round>();
         Game gameToDelete = service.getGameById(id);
-        RoundList = service.getRoundsForGame(gameToDelete);
+        if (gameToDelete == null){
+           return new ResponseEntity(HttpStatus.NOT_FOUND); 
+        }
+        List<Round> RoundList = service.getRoundsForGame(gameToDelete);
         if (RoundList == null) {
             service.deleteGameById(id);
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, "Id not found"
-            );
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Id not found");
         }
         for (Round r : RoundList) {
             service.deleteRoundById(r.getId());
